requirements.txt
construct>=2.9.29
pyasn1
asn1
termcolor
click
coloredlogs
IPython
bpylist2>=4.0.1
pygments
hexdump
arrow
daemonize
gpxpy
pykdebugparser>=1.1.5
pyusb>=1.2.1
tqdm
cached-property
requests
cmd2
packaging
pygnuutils>=0.0.4
cryptography>=35.0.0
pycrashreport>=0.0.5
fastapi[all]
uvicorn>=0.15.0
starlette
wsproto
nest_asyncio
Pillow
